Cryptol.Compiler.Cry2IR.Compile

Documentation

compileModule :: Module -> CryC ()

compileDeclGroup :: DeclGroup -> CryC ()

compileTopDecl :: Decl -> CryC ()

prepExprDecl :: Expr -> ([TParam], [(Name, Type)], Expr)

Identify expressions that are functions

seqElementType :: Type -> Type

seqLength :: Type -> StreamSize

compileExpr :: Expr -> Type -> ConvertM Expr

compileLam :: [(Name, Type)] -> Expr -> [Expr] -> Type -> ConvertM Expr

compileComprehension :: Type -> Type -> Expr -> [[Match]] -> ConvertM Expr

compileLocalDeclGroups :: [DeclGroup] -> ConvertM Expr -> ConvertM Expr

compileLocalDeclGroup :: DeclGroup -> ConvertM Expr -> ConvertM Expr

compileVar :: Name -> [Type] -> [Expr] -> Type -> ConvertM Expr

compileCall

Arguments

:: Name
Function name

-> [Type]
Type arguments at instantiation

-> [Expr]
Value arguments provided

-> Type
Wanted result type

-> ConvertM Expr

Compile a all to a function.

coerceTo :: Expr -> Type -> Expr

compileMapStream :: StreamSize -> Type -> Expr -> Name -> Expr -> ConvertM Expr

compileRecStream

Arguments

:: Name
Recursive stream name

-> Expr
Definition

-> ConvertM Expr

data CryRecEqn

Constructors

CryRecEqn

Fields

ceqName :: Name

ceqTy :: Type

ceqLen :: Type

ceqElTy :: Type

ceqDef :: Expr

isRecValueGroup :: [Decl] -> Either Doc [CryRecEqn]

Check if a recursive group is suitable for recursive stream compilation. Currently we require that all definitions are monomorphic streams, which should be the common case for locals (there can be still polymorphism from the enclosing declaration).

compileRecursiveStreams :: [CryRecEqn] -> ConvertM Expr -> ConvertM Expr