Cryptol.Compiler.Cry2IR.ConvertM

Documentation

newtype ConvertM a

Monad used to translate Cryptol expressions to IR

Constructors

ConvertM (SpecImpl a)

type SpecImpl = WithBase CryC '[ReaderT RO]

This is the implementation of the monad

data RO

Constructors

RO

Fields

roNumericParams :: Map TParam StreamSize

Information about type parameters. These will be either the variable used to represent them, or a constant, if we are working on an instantiation where this parameter is fixed.

roLocalIRNames :: Map NameId Name

Maps Cryptol name to an IR name, which has the IRType of the local

runConvertM :: ConvertM a -> CryC a

doIO :: IO a -> ConvertM a

Do some IO.

doCryC :: CryC a -> ConvertM a

Do CryC stuff.

doCryCWith :: (forall a. CryC a -> CryC a) -> ConvertM b -> ConvertM b

Do some nested CryC stuff.

lookupNumericTParam :: TParam -> ConvertM StreamSize

Get the value of a numeric type parameter.

unsupported :: Doc -> ConvertM a

Abort: we found something that's unsupported.

enterLoc :: Loc -> ConvertM a -> ConvertM a

enterFun :: Name -> FunInstance -> ConvertM a -> ConvertM a

enterLocal :: Name -> ConvertM a -> ConvertM a

withNumericTParams :: [(TParam, StreamSize)] -> ConvertM a -> ConvertM a

withLocals :: [(Name, Type)] -> ConvertM a -> ConvertM a

withIRLocals :: [Name] -> ConvertM a -> ConvertM a

Add some locals for the duration of a compiler computation

getLocal :: NameId -> ConvertM (Maybe Name)