Cryptol.Compiler.Cry2IR.Monad

Documentation

data SpecM a

runSpecM :: SpecM a -> CryC [a]

Run a computation with no constraints or parameters

Lifting

doIO :: IO a -> SpecM a

Do some IO.

doCryC :: CryC a -> SpecM a

Do CryC stuff.

doCryCWith :: (forall a. CryC a -> CryC a) -> SpecM b -> SpecM b

Errors

unsupported :: Text -> SpecM a

Abort: we found something that's unsupported.

panic :: HasCallStack => String -> [String] -> a

Solver

checkPossible :: SpecM ()

Check if the current set of properties are consistent. Kill the thread if not.

Constraints

data BoolInfo

Constructors

Known Bool
The parameter is/isn't bool on this path.

Unknown [Prop] [Trait]
If parameter is bool, then also add props. - If it is *not* bool, then add the given traits

data SizeConstraint

Constructors

IsFin

IsFinSize

IsInf

addTParams :: [TParam] -> SpecM ()

This should be done before adding constraints about the parameters.

addTrait :: Trait -> SpecM ()

addIsBoolProp :: TParam -> BoolInfo -> SpecM ()

Add a constraint on a (value) type parameter.

addNumProps :: [Prop] -> SpecM ()

Add some numeric properties to the current thread.

caseSize :: Type -> SpecM SizeConstraint

Returns the class for this type. May split the world.

maxSizeVal :: Integer

Largest value that will fit in a "size".

caseConst :: Type -> Ordering -> Integer -> SpecM Bool

caseIsInf :: Type -> SpecM Bool

Returns True if the given type is inf. May split the world.

caseBool :: TParam -> SpecM Bool

Returns if this type parameter is Bool. May split the world.

Locals

withLocals :: [(Name, Type)] -> SpecM a -> SpecM a

withIRLocals :: [Name] -> SpecM a -> SpecM a

Add some locals for the duration of a compiler computation

getLocal :: NameId -> SpecM (Maybe Name)

Queries

getTParams :: SpecM [TParam]

getTraits :: SpecM [Trait]

getBoolConstraint :: TParam -> SpecM BoolInfo

Get known constraints on a specific vairable.

zonk :: (ApSubst a, TName a ~ TParam) => a -> SpecM a

getSubst :: SpecM Subst

Find type parameters that can have only a single value.

checkSingleValue :: TParam -> SpecM (Maybe Nat')

Check if the given variable can have only one possible value.