Cryptol.Compiler.Cry2IR.Type

Documentation

compileValType :: Type -> ConvertM Type

This will only use a stream for infinite ones.

compileSizeType :: Type -> ConvertM Size

Compile a known finite Cryptol size type to an IR type.

compileStreamSizeType :: Type -> ConvertM StreamSize

Compile a Cryptol size type to an IR type.