Cryptol.Compiler.Driver

Documentation

cry2rust :: [FilePath] -> CryC ()

Load some Cryptol specs and generate Rust for them.

cryModRoots :: Set Text -> Module -> [Name]

Get public roots for this module, if any.

cryFilterDecls :: Set Text -> [Module] -> [Module]

Only keep declarations relevant for the given entry modules.

loadInputs :: [FilePath] -> CryC ()

Load Cryptol specs

cry2IR :: Module -> CryC [FunDecl]

Compile a Cryptol module to IR. The state of the monad is updated to keep track of the instances we generated.

ir2Rust :: ModName -> [FunDecl] -> CryC ()

Compile some IR to Rust, saving things in the file system. Modifies the state to track information about compiled modules.