Cryptol.Compiler.IR

Description

Basic IR for code generation.

Functions

data IRFunDecl tname name

A function declaration.

Constructors

IRFunDecl

Fields

irfName :: IRFunName name

irfType :: IRFunType tname

irfDef :: IRFunDef tname name

data IRFunDef tname name

The definition of a top-level function.

Constructors

IRFunPrim
A primitive function.

IRFunDef [name] (IRExpr tname name)
Function with definition. Includes names of the function parameters. Their types are in the function's type.

data IRFunName name

Function name, paired with an instance.

Constructors

IRFunName

Fields

irfnName :: IRFunNameFlavor name

irfnInstance :: FunInstance

data IRFunNameFlavor name

Various types of function names

Constructors

IRPrimName IRPrim
An IR primtive

IRDeclaredFunName name
A declared function

data IRPrim

This is for primitivies specifi to the IR (i.e., *not* the primitives of Cryptol)

Constructors

CryPrim PrimIdent
A Cryptol primitve

ArrayLit

ArrayLookup
(index : usize, xs: Array n T) -> T The index to look at is in the *size* args of the call The parameter is the name of the array Unspecified behavior if index >= n

WordLookup
(index: usize, xs: Word n) -> Bool

StreamLookup
{a,i} Integral i => (xs: Steram n a, index: i) -> a

StreamLookupBack
{a,i} Integral i => (xs: Steram n a, index: i) -> a

Tuple
Make a tuple.

TupleSel Int Int
select_n_of_N(tuple) Index n-th element in a tuple of size N

EqSize
Check if two sizes are the same. They are in the size args of the call

LtSize
Compare two size types. They are in the size args of the call

LeqSize
Compare two size types. They are in the size args of the call

FlatMap
(xs : Stream m a, f : a -> Stream n b) -> Stream (m*n) b

Zip
(xs : Stream m a, ys : Stream n b) -> Stream (min m n) (a,b)

ArrayToWord
Array n Bool  -> Word n

ArrayToStream
Array n a     -> Stream n a

WordToStream
Array n a     -> Stream n Bool

StreamToWord
Stream n Bool -> Word n

StreamToArray
Stream n a    -> Array n a

Head
Next element of external stream. (xs: Stream n a) -> a

Hist
Past element of a recursive stream. 0 is most recent. (index: usize) -> a

newtype FunInstance

A specific instance of a Cryptol function.

Constructors

FunInstance [ParamInfo]

data ParamInfo

A pattern in a function instance. NOTE: Ordering of constructors is important. More specific ones should come *before* less specific ones

Constructors

NumFixed Nat'

NumVar SizeVarSize

TyBool

TyNotBool

TyAny

Expressions

newtype IRExpr tname name

Expressions

Constructors

IRExpr (IRExprF tname name (IRExpr tname name))

data IRExprF tname name expr

The various flavours of expressions. Expressions are split in two parts because this makes it easier to write generic traversals.

Constructors

IRVar (IRName tname name)

IRCallFun (IRCall tname name expr)
Call a function

IRClosure (IRCall tname name expr)
Partial application

IRLam [IRName tname name] expr
Lambda

IRIf expr expr expr

IRLet (IRName tname name) expr expr

IRStream (IRStreamExpr tname name expr)

Local Names

data IRName tname name

Typed names. When compared, we only consider the name, not the type.

Constructors

IRName

Fields

irNameName :: name

irNameType :: IRType tname

Function Calls

data IRCall tname name expr

Information about a function call or a closure. Note: we keep the types of arguments around, as they are convenient to have for backends that use more refined representations of types, and so the argument types need to be adjusted.

Constructors

IRCall

Fields

ircFun :: IRCallable tname name expr

What we are calling

ircFunType :: !(IRFunType tname)

Type of the function we are calling (uninstantiated)

ircArgTypes :: [IRType tname]

Types of arguments (types of ircArgs)

ircResType :: IRType tname

Result of function, or closure type

ircArgs :: [expr]

Available arguments

data IRCallable tname name expr

Something that can be called.

Constructors

IRTopFun (IRTopFunCall tname name)
A top level declaration

IRFunVal expr
A closure value

data IRTopFunCall tname name

A call to a top level function, containing common parameters that are always present.

Constructors

IRTopFunCall

Fields

irtfName :: IRFunName name

irtfTypeArgs :: [IRType tname]

Type arguments

irtfSizeArgs :: [(IRSize tname, SizeVarSize)]

Size arguments

topCallTypeArgs :: IRTopFunCall tname name -> [IRType tname]

callPrimG

Arguments

:: IRPrim
Primitive name

-> [(IRSize tname, SizeVarSize)]
Size arguments

-> [IRExpr tname name]
Expression arguments

-> IRType tname
Result type

-> IRExpr tname name

Call a primitive (no instances, or type arguments).

callSizePrim :: IRPrim -> [(IRSize tname, SizeVarSize)] -> IRType tname -> IRExpr tname name

callPrim :: IRPrim -> [IRExpr tname name] -> IRType tname -> IRExpr tname name

mkTuple :: [IRExpr tname name] -> IRExpr tname name

Streams

data IRStreamExpr tname name expr

Constructors

IRStreamExpr

Fields

irsType :: IRType tname

TStream l a

irsExterns :: [(IRName tname name, expr)]

These are external streams we traverse. They should only be used *uniquely*, so we can mutate them on use.

irsRec :: IRStreamRec expr

irsNext :: expr

a

data IRStreamRec expr

Constructors

NonRecStream

RecStream expr
TArray hist a, initializer

Types

data IRType tname

Value types.

Constructors

TBool
Boolean

TInteger
Unbounded integer

TIntegerMod (IRSize tname)
Z

TRational
Rational number

TFloat
Floating point small

TDouble
Floating point large

TWord (IRSize tname)
Bit vector

TArray (IRSize tname) (IRType tname)
Array

TStream (IRStreamSize tname) (IRType tname)
Iterator

TTuple [IRType tname]
Tuple

TFun [IRType tname] (IRType tname)
Function types

TPoly tname
Polymorphic

class HasType t where

Things that have a type.

Methods

typeOf :: t -> IRType (TName t)

Compute the type of something.

Size Types

data IRStreamSize tname

Size types that could be infinite.

Constructors

IRInfSize
Infinite size

IRSize (IRSize tname)
Finite size

data IRSize tname

Size "types". These are typically not erased, to they are really values. These are supposed to be finite.

Constructors

IRFixedSize Integer
A specific size

IRPolySize (IRSizeName tname)
Polymorphic size; finite

IRComputedSize TFun [IRSize tname]
Computed size

data IRSizeName tname

The name of a size variable.

Constructors

IRSizeName

Fields

irsName :: tname

irsSize :: SizeVarSize

data SizeVarSize

The size of a numeric type parameter.

Constructors

MemSize
value will fit in usize

LargeSize
value *is* large, use BigInt

isKnownStreamSize :: IRStreamSize tname -> Maybe Nat'

Is this a known size type or infinity.

isKnownSize :: IRSize tname -> Maybe Integer

Is this a known size type.

pattern K :: Integer -> IRStreamSize tname

maxSizeVal :: Integer

Largest value that will fit in a "size".

Function Types

data IRFunType tname

The type of a function.

Constructors

IRFunType

Fields

ftTypeParams :: [tname]

Value type parameters.

ftTraits :: [IRTrait tname]

Constraints on the type paramters.

ftSizeParams :: [IRSizeName tname]

Size parameters.

ftParams :: [IRType tname]

Normal parameters.

ftResult :: IRType tname

Function result.

Traits

data IRTrait tname

A constraint on a type parameter.

Constructors

IRTrait

Fields

irTraitName :: IRTraitName

irTraitType :: tname

data IRTraitName

Names of constraints (aka traits in Rust, classes in Haskell).

traitMap :: Ord tname => [IRTrait tname] -> Map tname [IRTraitName]

Group traits by their variable name.

Type functions

type family VName a

type family TName t