Cryptol.Compiler.IR.Prims

Documentation

data IRPrim

This is for primitivies specifi to the IR (i.e., *not* the primitives of Cryptol)

Constructors

CryPrim PrimIdent
A Cryptol primitve

ArrayLit

ArrayLookup
(index : usize, xs: Array n T) -> T The index to look at is in the *size* args of the call The parameter is the name of the array Unspecified behavior if index >= n

WordLookup
(index: usize, xs: Word n) -> Bool

StreamLookup
{a,i} Integral i => (xs: Steram n a, index: i) -> a

StreamLookupBack
{a,i} Integral i => (xs: Steram n a, index: i) -> a

Tuple
Make a tuple.

TupleSel Int Int
select_n_of_N(tuple) Index n-th element in a tuple of size N

EqSize
Check if two sizes are the same. They are in the size args of the call

LtSize
Compare two size types. They are in the size args of the call

LeqSize
Compare two size types. They are in the size args of the call

FlatMap
(xs : Stream m a, f : a -> Stream n b) -> Stream (m*n) b

Zip
(xs : Stream m a, ys : Stream n b) -> Stream (min m n) (a,b)

ArrayToWord
Array n Bool  -> Word n

ArrayToStream
Array n a     -> Stream n a

WordToStream
Array n a     -> Stream n Bool

StreamToWord
Stream n Bool -> Word n

StreamToArray
Stream n a    -> Array n a

Head
Next element of external stream. (xs: Stream n a) -> a

Hist
Past element of a recursive stream. 0 is most recent. (index: usize) -> a