Cryptol.Compiler.Rust.CompileStream

Documentation

type CompExpr = ?genExpr :: ExprContext -> Expr -> Rust (PartialBlock RustExpr)

genStream :: CompExpr => StreamExpr -> Rust (PartialBlock RustExpr)

data StoredStream

Information about a stream stored in a stream closure.

Constructors

StoredStream

Fields

stoName :: Name

Original name of the stored stream

stoDef :: Maybe Expr

Definition for stream, if mutable

stoStrLabel :: Ident

Value name of the stream

stoStrTyParam :: Ident

The type parameter for it's type

stoStrType :: RustType

The type parameter as a type

stoStrTraits :: [RustPath]

Constraints on the type param

streamTypeVar :: Int -> (Name, Maybe Expr) -> Rust StoredStream

Information about the name and type of an external stream vairable Note that this adds the stream variable to the current scope, and we should remove it at the end.

makeGenerics :: [TParam] -> [StoredStream] -> Rust [(Ident, [RustPath])]

Make type parameters and their constraints

streamMacro :: [(Ident, [RustPath])] -> RustType -> [(Ident, RustType, RustExpr)] -> Maybe (Integer, RustExpr) -> RustExpr -> RustExpr