Cryptol.Compiler.Rust.CompileTrait

Documentation

traitNeedsLen :: Trait -> Set TParam

lenParamType :: TParam -> Rust RustType

isCryType :: TParam -> Rust RustWherePredicate

compileTrait :: Trait -> Rust (Set TParam, RustWherePredicate)

Compile a trait. The set returns type parameters that need an associated Length parmaeter.

lenParamFor :: Type -> Rust RustExpr

Given a type, compute the parameter for the Length trait in the Rust run-time system.