Cryptol.Compiler.Rust.CompileType

Description

Translation of names from IR to Rust.

Documentation

data TypeMode

Different representations for a Cryptol type.

Constructors

AsArg (Maybe RustLifetime)
Borrowed represewntation, used to pass arguments

AsOwned
Owned representation, currently used for locals, function results, and when the type is a field of another type

data TypeUse

Type use XXX: A lot of these complications are because of streams and functions, which should probably be treated specially...

Constructors

TypeInFunSig
Argument or result of a functoin

TypeAsParam
Type application in a function call

TypeAsStored
Stored in another type (e.g., array/stream)

compileType :: TypeUse -> TypeMode -> Type -> Rust RustType

Compute the Rust type used to represent the given Cryptol type.

byRef :: TypeMode -> RustType -> RustType

streamTraitPath :: RustType -> RustPath

cryIntegerType :: RustType

cryZType :: RustType

cryRationalType :: RustType

cryVectorType :: RustType -> RustType

cryDWordType :: RustType

cryDWordRefType :: Maybe RustLifetime -> RustType