Cryptol.Compiler.Rust.Crate

Description

Utilities for generating crates and including the generated runtime libraries

Documentation

mkCrate

Arguments

:: Bool
Do we want an example driver

-> String
Name of the crate

-> FilePath
Directory to store the crate in

-> [Ident]
Crate `lib.rs` with these modules, if non-empty

-> IO ()

Make a barebones crate with the given name and the given directory.