Cryptol.Compiler.Rust.Monad

Documentation

data GenInfo

Constructors

GenInfo

Fields

genCurModule :: ModName

genExternalModules :: Map ModName ExtModule

runRustM :: GenInfo -> Rust a -> IO a

data ExprContext

Constructors

OwnContext
We are generating an ownded expression

BorrowContext
We are generating a borrowed expression

MutContext
At the moment, this is only used for Head, and is extremely special cased

ownIfStream :: Type -> ExprContext

Use Own for streams and functions, and Borrow otherwise

type RustImpl = WithBase IO [ReaderT RO, StateT RW]

newtype Rust a

Constructors

Rust (RustImpl a)

doIO :: IO a -> Rust a

Run an IO computation in the Rust monad

data ExtModule

Information about previously compiled modules.

Constructors

ExtModule

Fields

extModuleName :: [Ident]

Name of module

extModuleNames :: Map FunName Ident

Functions defined in the module

data RO

Constructors

RO

Fields

roModName :: ModName

The current module we are working on

roExternalNames :: Map ModName ExtModule

Names defined in different modules. Read only.

roLoc :: Loc

Location, for error reporting

data RW

Constructors

RW

Fields

rwLocalFunNames :: NameMap FunName

Names in the current module

rwLocalNames :: LocalNames

Names in the current function

resetFunLocalNames :: Rust ()

Reset names local to a function.

bindLocal

Arguments

:: (a -> LocalNames -> (Ident, LocalNames))
How to bind it

-> a
Name of the local thing

-> Rust Ident

Bind a local names

bindLocalLet :: NameId -> (Ident -> Rust a) -> Rust a

Bind a local for the duraiton of the given computation

removeLocalLet :: NameId -> Rust ()

In some cases, instead of using bindLocalLet it is more convenient to first add some locals and them remove them. This avoids having to nest stuff.

addTypeBound :: TParam -> RustWherePredicate -> Rust ()

Add a type bound for a parameter.

getTypeBounds :: TParam -> Rust [RustWherePredicate]

Get the type bounds for this parameter.

withSameCont :: [Rust a] -> Rust [a]

Execute the given computation with the same continuation. This means that all expressions will see the same set of used variables. The resulting expressions uses a variable if any of the computation does. (e.g., then "then" and "else" branches of if-then-else)

getUsed :: Rust (Set Ident)

setUsed :: Set Ident -> Rust ()

inStreamClosure :: ExprContext -> Rust a -> Rust ([(Ident, TParam)], [(Ident, SizeVarSize)], a)

Do something without affecting the continuation, and variables are only accessed through Self.

isInStreamClosure :: Rust (Maybe StreamClosureInfo)

Are we in a strem closure?

mapStreamClosure :: (StreamClosureInfo -> StreamClosureInfo) -> Rust Bool

bindFun :: FunName -> Rust Ident

Bind a function in this module

getTParams :: Rust (TParam -> RustType)

lookupTParam :: TParam -> Rust Ident

Get the identifiier corresponding to a type parameter.

lookupSizeParam :: SizeName -> Rust RustExpr

Get the expresssion for a size parameter.

lookupNameId :: NameId -> Rust (Bool, Bool, RustExpr)

Get the identfier for a name. Returns (isThisLocal?, isThisLastUse?,Expr) isThisLocal is True if this is `let` bound local variable isThisLastUse is True if the variable is not mentioned in the continuation.

lookupNameRustIdent :: NameId -> Rust Ident

Get the identifier for to use for the given local names

lookupLenParam :: TParam -> Rust RustExpr

Get the expression for a length parametes associated with the given type parameter.

isFunNameLocal :: FunName -> Rust Bool

Is this name in the module that we are currently compiling.

lookupFunName :: FunName -> Rust (Either IRPrim RustPath)

Get an expression corresponding to a named function

localScope :: Rust a -> Rust a

Evaluate a computation, forgetting any locally bound variables afterward.

getFunNames :: Rust (Map FunName Ident)

Get the ampping between IR names and Rust names for this module.

data LocalNames

Names local to a declaration

Constructors

LocalNames

Fields

lTypeNames :: NameMap TParam

Names for type params

lTypeBounds :: Map TParam [RustWherePredicate]

Bounds on type parameters we have. We keep this around, so that if we need to make a stream closure, we can copy them in the struct declaration.

lValNames :: NameMap NameId

Names for local values

lLenParams :: Map TParam Ident

Names for Length params. When generating these we use the "used" set of lValNames, because they are values.

lSizeParams :: Map TParam Ident

Names for size parameters. When generating these we use the "used" set of lValNames, because they are values.

lLocalVars :: Map Ident LocVarLocation

Names of local variable (i.e., not arguments)

lUsedVars :: Set Ident

Variables used by the current continuation. We generate code "backwards" (continuation first), and this field tells us which variables were mentioned in the continuation.

lInStreamClosure :: !(Maybe StreamClosureInfo)

If this is Just, then we are compiling a stream. * Then we should access local variables through self. * The ExprContext indicates how variables are stored in the closure (i.e., are they borrowed or owned by the closure).

data StreamClosureInfo

Constructors

StreamClosureInfo

Fields

cloLenIdents :: Map Ident TParam

Lenght variables that need to be stored in the closure

cloSizeIdents :: Map Ident SizeVarSize

Size variables that need to be stored in the closure

cloContext :: ExprContext

data LocVarLocation

How to access a local variable

Constructors

KnownLocal
Always access as a local variable

LocalOrClosure
Acces through closure, if in a closure context.

emptyLocalNames :: LocalNames

Empty local names, useful when starting a new declaration.

addLocalType :: TParam -> LocalNames -> (Ident, LocalNames)

Bind a type parameter.

addLocalVar :: Maybe LocVarLocation -> NameId -> LocalNames -> (Ident, LocalNames)

Bind a local variable/paramter.

removeLocalVar :: NameId -> Ident -> LocalNames -> LocalNames

addLocalLenghtParam :: TParam -> LocalNames -> (Ident, LocalNames)

Bind a local parameter used for the Length trait.

addLocalSizeParam :: TParam -> LocalNames -> (Ident, LocalNames)

Bind a local parameter used for the Length trait.

enter :: Doc -> Rust a -> Rust a

Enter a named scope (e.g., a delcaration)

unsupported :: Doc -> Rust a

Raise an error, that something is not yet supported