Cryptol.Compiler.Rust.NameMap

Description

Maps names to rust identifier, and avoids reusing identifier.

Documentation

data NameMap a

Associate names with Rust identifiers, and keeps track of which rust identifiers we've already used.

Constructors

NameMap

Fields

lUsed :: Set Ident

lMap :: Map a Ident

emptyNameMap :: Ord a => NameMap a

An empty mpa.

addName :: (Ord a, RustIdent a) => a -> NameMap a -> (Ident, NameMap a)

Pick a Rust name for something, ensuring that it does not clash with any previously used names.

lookupName :: (PP a, Ord a) => a -> NameMap a -> Ident

Lookup a name. Panics if not defined.