Cryptol.Compiler.Rust.Names

Description

Translation of names from IR to Rust.

Documentation

class RustIdent a where

Things that provide names.

Methods

rustIdent :: a -> [Ident]

Pick a Rust identifier for the given thing string. We compute a non-empty list of options, where the identifiers earlier on are to be preferred.

newtype TraitLengthName

Name of the length parameter for the methods that require a dynamic length.

Constructors

TraitLengthName TParam

newtype SizeParamName

A name of a size parameter, don't use camel case.

Constructors

SizeParamName TParam

rustIdentAvoiding :: (String -> String) -> Set Ident -> [Ident] -> Ident

Pick a name for something, so that it does not clash with the given names.

changeIdent :: (String -> String) -> Ident -> Ident

Change a rust identifier. Preserves the rawness.

snakeCase :: String -> String

Use snake_case

screamingSnakeCase :: String -> String

Use SCREAMING_SNAKE_CASE

upperCamelCase :: String -> String

Use UpperCamelCase

rustModNameChunks :: ModName -> [Ident]

Commonly used names

cryptolCrateString :: String

Qualifier to use when accessing RTS names

cryptolCrate :: Ident

Qualifier to use when accessing RTS names