Cryptol.Compiler.Rust.Utils

Description

Utilities for generating Rust code Nothing here should be specific to the compiler.

Documentation

type RustSourceFile = SourceFile ()

type RustAttribute = Attribute ()

type RustPath = Path ()

type RustType = Ty ()

type RustExpr = Expr ()

type RustStmt = Stmt ()

type RustBlock = Block ()

type RustItem = Item ()

type RustImplItem = ImplItem ()

type RustMethodSig = MethodSig ()

type RustArg = Arg ()

type RustGenerics = Generics ()

type RustLifetime = Lifetime ()

type RustLifetimeDef = LifetimeDef ()

type RustTyParam = TyParam ()

type RustWherePredicate = WherePredicate ()

type RustWhereClause = WhereClause ()

type RustLit = Lit ()

type PartialBlock a = ([RustStmt], a)

A block under construction

justExpr :: a -> PartialBlock a

Make a partial block with no statememtns.

finishBlock :: PartialBlock RustExpr -> RustBlock

Turn a partial block into an actual block.

dummySpan :: Span

simplePath :: Ident -> RustPath

simplePath' :: [Ident] -> RustPath

pathWithTypes :: [Ident] -> [RustType] -> RustPath

pathAddTypeSuffix :: RustPath -> [RustType] -> RustPath

pathAddLifetimeSuffix :: RustPath -> [RustLifetime] -> RustPath

fnTraitPath :: [RustType] -> RustType -> RustPath

Fn((u8) -> u8)

iteratorWithTypePath :: RustType -> RustPath

IteratorItem=T

lifetime :: Name -> RustLifetime

lifetimeMaybe :: Maybe RustLifetime -> RustLifetime

typeQualifiedExpr :: RustType -> RustPath -> RustExpr

pathExpr :: RustPath -> RustExpr

block :: [RustStmt] -> RustBlock

block' :: [RustStmt] -> RustExpr -> RustBlock

stmtNoSemi :: RustExpr -> RustStmt

assign :: RustExpr -> RustExpr -> RustStmt

localLet :: Ident -> Maybe RustType -> RustExpr -> RustStmt

mkRustCall :: RustExpr -> [RustExpr] -> RustExpr

callMethod :: RustExpr -> Ident -> [RustExpr] -> RustExpr

callMacro :: RustPath -> [RustExpr] -> RustExpr

callMacro' :: RustPath -> TokenStream -> RustExpr

mkClosure :: [(Ident, RustType)] -> [RustStmt] -> RustExpr -> RustExpr

mkGenerics :: [RustLifetimeDef] -> [RustTyParam] -> RustWhereClause -> RustGenerics

noGenerics :: RustGenerics

blockExprIfNeeded :: [RustStmt] -> RustExpr -> RustExpr

identPat :: Ident -> Pat ()

todoExp :: String -> RustExpr

todoBlock :: String -> RustBlock

exprStmt :: RustExpr -> RustStmt

addrOf :: RustExpr -> RustExpr

rustArray :: [RustExpr] -> RustExpr

rustIf :: RustExpr -> RustBlock -> RustBlock -> RustExpr

rustTry :: RustExpr -> RustExpr

sourceFile :: Maybe Name -> [RustAttribute] -> [RustItem] -> RustSourceFile

disableWarning :: Ident -> RustAttribute

mkFnItem :: Ident -> RustGenerics -> [(Ident, RustType)] -> RustType -> RustBlock -> RustItem

mkFnItem_ :: Ident -> RustGenerics -> [(Ident, RustType)] -> RustBlock -> RustItem

mkUse :: [Ident] -> RustItem

Make a simple use path

mkUseGlob :: [Ident] -> RustItem

Make a simple use path

pubMod :: Ident -> RustItem

Declare a public module.

mkImplTrait :: RustGenerics -> RustPath -> RustType -> [RustImplItem] -> RustItem

implType :: Ident -> RustType -> RustImplItem

Definition of an associated type in a trait impl

implMethod :: Ident -> RustGenerics -> RustMethodSig -> RustBlock -> RustImplItem

methodSig :: [RustArg] -> Maybe RustType -> RustMethodSig

selfRef :: Mutability -> RustArg

structField :: Ident -> RustType -> StructField ()

mkIntLit :: Suffix -> Integer -> RustLit

mkU8Lit :: Integer -> RustLit

mkU64Lit :: Integer -> RustLit

mkUSizeLit :: Integer -> RustLit

boolLit :: Bool -> RustLit

strLit :: Text -> RustLit

litExpr :: RustLit -> RustExpr

unitExpr :: RustExpr

tupleExpr :: [RustExpr] -> RustExpr

tupleSelect :: RustExpr -> Int -> RustExpr

fieldSelect :: RustExpr -> Ident -> RustExpr

arrayExpr :: [RustExpr] -> RustExpr

indexExpr :: RustExpr -> RustExpr -> RustExpr

binExpr :: BinOp -> RustExpr -> RustExpr -> RustExpr

inferType :: RustType

simpleType :: String -> RustType

typeQualifiedType :: RustType -> RustPath -> RustType

pathType :: RustPath -> RustType

unitType :: RustType

tupleType :: [RustType] -> RustType

boolType :: RustType

vectorOfType :: RustType -> RustType

fixedArrayOfType :: RustType -> Integer -> RustType

sliceType :: RustType -> RustType

refType :: Maybe RustLifetime -> RustType -> RustType

mutRefType :: RustType -> RustType

implTraitType :: [RustPath] -> RustType

Make a type like `impl T` where T is a trait.

implFnTraitType :: [RustType] -> RustType -> RustType

tyParam :: Ident -> RustTyParam