Cryptol.IR.LambdaLift

Description

Lift local functions to the top level. 1. Free local variables become parameters 2. Any type variables that are mentioned become type parameters 3. Any constraints that mention these variables. 3.5 Note that if these constraints mention variables that are not mentioned anywhere else we need to also add *them*, which in turn might add more constraints, etc. 4. Each use of the local function needs to pass these additional parameters. 5. Note that locally defined functions (i.e., non-parameter ones) do not become paramters because they will be lifter.

In summary, given: `e where ds` and the ds are in dependency order. For each d in order: * rewrite definition to eliminte any functions that were already lifted. * compute free variables (types, values, predicates) * using this add a new top-level declaration. * remember how to change the instantiations, in the continuation Once we've done all the ds we need to rewrite the e

f ts ps xs ~> f (freeTs++ts) (newPs ++ ps) (freeVars ++ xs)

In a smillar way we also need to lift lambdas, except that they are only goind to be used in place.

When we are a lifting a polymorphic local, we should make sure to not add constraints that the function already has, so in the above example, the (++) in the constratins is a bit more complicated. Similarly, in the definition of the lifted function we have to insert the correct number of PAbs.

Local Polymorphic Values

Note that we also lambda lift local polymorphic values, which hopefully should not be very common as we have local mono binds on by default.

Documentation

llModule :: Map Name Schema -> Supply -> ModuleG mname -> ((ModuleG mname, Map Name Schema), Supply)