Cryptol.Compiler.IR

Description

Basic IR for code generation.

Documentation

data IRFunName name

Function name, paired with an instance.

Constructors

IRFunName

Fields

irfnName :: IRFunNameFlavor name

irfnInstance :: FunInstance

data IRFunNameFlavor name

Various types of function names

Constructors

IRPrimName IRPrim
An IR primtive

IRDeclaredFunName name
A declared function

data IRName tname name

Typed names. When compared, we only consider the name, not the type.

Constructors

IRName

Fields

irNameName :: name

irNameType :: IRType tname

data IRFunDecl tname name

A function declaration

Constructors

IRFunDecl

Fields

irfName :: IRFunName name

irfType :: IRFunType tname

irfDef :: IRFunDef tname name

data IRFunDef tname name

The definition of a top-level function.

Constructors

IRFunPrim

IRFunDef [name] (IRExpr tname name)

newtype IRExpr tname name

Expressions

Constructors

IRExpr (IRExprF tname name (IRExpr tname name))

data IRExprF tname name expr

The various flavours of expressions. Expressions are split in two parts because this makes it easier to write generic traversals.

Constructors

IRVar (IRName tname name)

IRCallFun (IRCall tname name expr)
Call a function

IRClosure (IRCall tname name expr)
Partial application

IRLam [IRName tname name] expr
Lambda

IRIf expr expr expr

IRLet (IRName tname name) expr expr

IRStream (IRStreamExpr tname name expr)

data IRCall tname name expr

Information about a function call or a closure. Note: we keep the types of arguments around, as they are convenient to have for backends that use more refined representations of types, and so the argument types need to be adjusted.

Constructors

IRCall

Fields

ircFun :: IRCallable tname name expr

What we are calling

ircArgTypes :: [IRType tname]

Types of arguments

ircResType :: IRType tname

Result of function, or closure type

ircArgs :: [expr]

Available arguments

data IRStreamExpr tname name expr

A group of stream equations.

Constructors

IRStreamExpr

Fields

irseCurIndex :: IRSizeName tname

Name used to refer to the current index. This becomes a parmater to the next function for the stream, and may be used in the body (e.g., to decide if we are in the base case or in the recursive case.

irseDecls :: [IRStreamDef tname name expr]

Defines how to compute the next element of a stream. An expression may use the values of any history variables, or of previously defined expressions in the list XXX: shall we allow the latter?.

This can be used to define the body of the next function.

irseEntries :: [IRName tname name]

Streams whose values are used outside the recursive group. Each of these needs its own copy of the *entire* steram state, as they may be consumed in different ways.

If this is a singleton, we treat the expression as a single stream. If there are more than one entries, then we generate a tuple.

data IRCallable tname name expr

Something that can be called.

Constructors

IRTopFun (IRTopFunCall tname name)
A top level declaration

IRFunVal expr
A closure value

data IRTopFunCall tname name

A call to a top level function, containing common parameters that are always present.

Constructors

IRTopFunCall

Fields

irtfName :: IRFunName name

irtfTypeArgs :: [IRType tname]

Type arguments

irtfSizeArgs :: [(IRSize tname, SizeVarSize)]

Size arguments

data IRStreamDef tname name expr

A stream equation, and information about its history.

Constructors

IRStreamDef

Fields

irsdName :: IRName tname name

irsdHistory :: IRStreamSize tname

irsdDef :: expr

type family VName a

class HasType t where

Things that have a type.

Methods

typeOf :: t -> IRType (TName t)

Compute the type of something.

ppIRTopFunCall :: (PP tname, PP name) => [Doc] -> IRTopFunCall tname name -> Doc

module Cryptol.Compiler.IR.Prims

module Cryptol.Compiler.IR.Type

module Cryptol.Compiler.IR.Common