Cryptol.Compiler.IR.Prims

Documentation

data IRPrim

This is for primitivies specifi to the IR (i.e., *not* the primitives of Cryptol)

Constructors

CryPrim PrimIdent
A Cryptol primitve

MakeSeq
Make a sequence literal. The arguments are the elementes of the sequence The result type in the call has the type of sequences we are making.

Tuple
Make a tuple.

TupleSel Int Int
select_n_of_N(tuple) Index n-th element in a tuple of size N

EqSize
Check if two sizes are the same. They are in the size args of the call

LeqSize
Compare two size types. They are in the size args of the call

Map
(xs : Stream n a, f : a -> b) -> Stream n b

FlatMap
(xs : Stream m a, f : a -> Stream n b) -> Stream (m*n) b

Zip
(xs : Stream m a, ys : Stream n b) -> Stream (min m n) (a,b)

Collect
(xs : Stream n a) -> Array n a or (xs : Stream n Bool) -> Word n

Iter
(xs : Array n a) -> Stream n a or (xs : Wor n) -> Stream n Bit

LookupHistory
The index to look at is in the *size* args of the call The parameter is the name of the history variable