Cryptol.Compiler.Rust.Monad

Documentation

data GenInfo

Constructors

GenInfo

Fields

genCurModule :: ModName

genExternalModules :: Map ModName ExtModule

runGenM :: GenInfo -> Gen a -> a

type GenM = WithBase Id [ReaderT RO, StateT RW]

newtype Gen a

Constructors

Gen (GenM a)

data ExtModule

Information about previously compile modules.

Constructors

ExtModule

Fields

extModuleName :: RustPath

Name of module

extModuleNames :: Map FunName Ident

Functions defined in the module

data RO

Constructors

RO

Fields

roModName :: ModName

The current module we are working on

roExternalNames :: Map ModName ExtModule

Names defined in different modules. Read only.

data RW

Constructors

RW

Fields

rwLocalFunNames :: NameMap FunName

Names in the current module

rwLocalNames :: LocalNames

Names in the current function

bindLocal

Arguments

:: (a -> LocalNames -> (Ident, LocalNames))
How to bind it

-> a
Name of the local thing

-> Gen Ident

Bind a local names

bindFun :: FunName -> Gen Ident

Bind a function in this module

getTParams :: Gen (TParam -> RustType)

lookupTParam :: TParam -> Gen RustType

Get the type corresponding to a type parameter.

lookupNameId :: NameId -> Gen RustExpr

Get the expresssion for a local.

lookupFunName :: FunName -> Gen (Either IRPrim RustExpr)

Get an expression corresponding to a named function

localScope :: Gen a -> Gen a

Evaluate a computation, forgetting any locally bound variables afterward.

data LocalNames

Names local to a declaration

Constructors

LocalNames

Fields

lTypeNames :: NameMap TParam

lValNames :: NameMap NameId

emptyLocalNames :: LocalNames

Empty local names, useful when starting a new declaration.

addLocalType :: TParam -> LocalNames -> (Ident, LocalNames)

Bind a type parameter.

addLocalVar :: NameId -> LocalNames -> (Ident, LocalNames)

Bind a local variable/paramter.